started extracing out views and models. started working on navigation link for each cell
